+++ R/swcGetData.R
@@ -1,7 +1,7 @@
 #' Import historic commune data from the web
 #' 
 #' This functon returns historic commune.  If necessary,
-#' \code{\link{read_historicized_commune_data}} is called.  The data is cached
+#' \code{\link{swcReadData}} is called.  The data is cached
 #' with the help of the \link{R.cache} package.
 #' 
 #' @return A named list with the components \code{canton}, \code{district} and
@@ -9,11 +9,11 @@
 #' 
 #' @param force If \code{TRUE}, reload data from the web.
 #' 
-#' @usage get_historicized_commune_data(force = FALSE)
+#' @usage swcGetData(force = FALSE)
 #' 
-#' @seealso \code{\link{read_historicized_commune_data}}
+#' @seealso \code{\link{swcReadData}}
 #' 
 #' @export
 #' @importFrom R.cache addMemoization
-#' @include read_historicized_commune_data.R
-get_historicized_commune_data <- R.cache::addMemoization(read_historicized_commune_data)
+#' @include swcReadData.R
+swcGetData <- R.cache::addMemoization(swcReadData)