+++ R/swcReadData.R
@@ -10,9 +10,9 @@
 #' 
 #' @export
 #' @importFrom logging logdebug
-read_historicized_commune_data <- function() .read_historicized_commune_data()
+swcReadData <- function() .swcReadData()
 
-.read_historicized_commune_data <- function() {
+.swcReadData <- function() {
   RECORD_HIST_URL <- 'http://www.bfs.admin.ch/bfs/portal/de/index/infothek/nomenklaturen/blank/blank/gem_liste/02.Document.96666.zip'
   zip.file.name <- tempfile(fileext='.zip')
   logdebug(zip.file.name)